static file to string
remove wrong LogicException
