+++ src/CodeGenMethod.php
@@ -4,7 +4,6 @@ declare(strict_types=1);
 
 namespace Ray\Aop;
 
-use _HumbugBox90c4dcb919ed\Symfony\Component\Console\Exception\LogicException;
 use Doctrine\Common\Annotations\AnnotationReader;
 use PhpParser\Builder\Method;
 use PhpParser\Builder\Param;
@@ -18,7 +17,6 @@ use PhpParser\NodeTraverser;
 use PhpParser\Parser;
 use PhpParser\PrettyPrinter\Standard;
 use Ray\Aop\Annotation\AbstractAssisted;
-use function is_string;
 
 final class CodeGenMethod
 {
@@ -171,12 +169,7 @@ final class CodeGenMethod
      */
     private function getTemplateMethodNodeStmts() : array
     {
-        $templateFile = dirname(__DIR__) . '/template/AopTemplate.php';
-        $code = file_get_contents($templateFile);
-        if (! is_string($code)) {
-            throw new LogicException; // @codeCoverageIgnore
-        }
-        /** @var string $code */
+        $code = $this->getTemplateCode();
         $node = $this->parser->parse($code)[0];
         if (! $node instanceof Class_) {
             throw new \LogicException; // @codeCoverageIgnore
@@ -187,5 +180,57 @@ final class CodeGenMethod
         }
 
         return $methodNode->stmts;
+    }
+
+    /**
+     * Return CodeGenTemplate string
+     *
+     * Compiler takes only the statements in the method. Then create new inherit code with interceptors.
+     *
+     * @see http://paul-m-jones.com/archives/182
+     * @see http://stackoverflow.com/questions/8343399/calling-a-function-with-explicit-parameters-vs-call-user-func-array
+     * @see http://stackoverflow.com/questions/1796100/what-is-faster-many-ifs-or-else-if
+     * @see http://stackoverflow.com/questions/2401478/why-is-faster-than-in-php
+     */
+    private function getTemplateCode() : string
+    {
+        return <<<'EOT'
+<?php
+class AopTemplate extends \Ray\Aop\FakeMock implements Ray\Aop\WeavedInterface
+{
+    /**
+     * @var array
+     *
+     * [$methodName => [$interceptorA[]][]
+     */
+    public $bindings;
+
+    /**
+     * @var bool
+     */
+    private $isIntercepting = true;
+
+    /**
+     * Method Template
+     *
+     * @param mixed $a
+     */
+    public function templateMethod($a, $b)
+    {
+        if ($this->isIntercepting === false) {
+            $this->isIntercepting = true;
+
+            return parent::templateMethod($a, $b);
+        }
+
+        $this->isIntercepting = false;
+        // invoke interceptor
+        $result = (new \Ray\Aop\ReflectiveMethodInvocation($this, __FUNCTION__, [$a, $b], $this->bindings[__FUNCTION__]))->proceed();
+        $this->isIntercepting = true;
+
+        return $result;
+    }
+}
+EOT;
     }
 }
